+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed effective radius in MieQuery.H to remove in place units scaling; made
   corresponding change in Chem_SettlingSimple
 - Updated optics lookup tables to accommodate area and effective radius calculation
-- Check userRC in ESMF_GridCompRun in GOCART2G gridcomp
+I- Check userRC in ESMF_GridCompRun in GOCART2G gridcomp
 - The pressure lid change associated with the introduction of run0 to set 0 above the lid
 - Fwet value in dust modified from 0.8 to 1.0
 - Dust and Sea salt Emission scale factors updated for L181
@@ -47,6 +47,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed erroneous/extraneous friendly attributes to internal state for DU and NI
   when in data_driven mode
 - typo in filepath for BR optics file
+- Units error for dust and sea salt radius in NIheterogenousChem in Process Library corrected.
+- stochiometric coefficent correction for heterogenous production diagnostic NI_phet
+  in NIheterogenousChem in Process Library.
+- corrected references for RH factor for hno3 uptake on dust in NIheterogenousChem in Process Library.
+- units error for external data hno3 corrected in NI instance file and in Process Library.
+- corrected units error for hno3 column diagnostic in Process Library.
+
+
 - Bug fix for "WetRemovalUFS"
 
 ### Added
@@ -56,6 +64,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   coupling of GMI OH, NO3, H2O2 to sulfur chemistry mechanism
 - Added a callback to allow a chemistry module to call for optical properties needed
   for photolysis calculation; currently used for GMI with CloudJ
+- Added 3d diagnostics for nitrate production.
+- Added logic in NI2G_GridCompMod.F90 through NI2G_instance_NI.rc to allow two-way
+  coupling of GMI HNO3 to nitrate chemistry mechanism.
 - Implementation of the "WetRemovalUFS" option for sulfate
 
 ## [v2.4.3] - 2025-07-21